Move whitelist host code to networkmanager
This means no :adblock-update after modifying the whitelist.
